+++ client/src/routes/privateRoutes.tsx
